+++ src/types/metadata.d.ts
@@ -1,0 +1,7 @@
+import { IAttr } from "./form";
+
+export interface IMetadata extends IAttr {
+    key: string;
+    _key: string;
+    value: any;
+}