+++ src/utils/export.ts
@@ -1,7 +1,10 @@
 import yaml from "js-yaml";
 import moment from "moment";
 
-import { IForm } from "./../types/form";
+import { tokenSplit } from "./string";
+
+import { IForm, IAttr } from "./../types/form";
+import { IMetadata } from "./../types/metadata";
 
 const KEYS_ORDER = [
     "title",
@@ -14,27 +17,10 @@ const KEYS_ORDER = [
     "memo",
 ];
 
-export function dump(form: IForm): string {
-    const attrs: Record<string, any> = {
-        title: form.basics.title,
-        id: form.others.id,
-        date: moment(form.basics.created, "YYYYMMDDHHmmss").toDate(),
-        lastmod: moment(form.basics.updated, "YYYYMMDDHHmmss").toDate(),
-        // date: moment(form.basics.created, "YYYYMMDDHHmmss").utc().format(),
-        // lastmod: moment(form.basics.updated, "YYYYMMDDHHmmss").utc().format(),
-    };
-
-    if (form.basics.name) attrs.name = form.basics.name;
-    if (form.basics.memo) attrs.memo = form.basics.memo;
-    if (form.basics.tags.length > 0) attrs.tags = form.basics.tags;
-    if (form.basics.alias.length > 0) attrs.alias = form.basics.alias;
-
-    form.customs.reduce((attrs, custom) => {
-        attrs[custom.key] = custom.value;
-        return attrs;
-    }, attrs);
-
-    return yaml.dump(attrs, {
+export function dump(
+    obj: any,
+    options: yaml.DumpOptions = {},
+    defaultOptions: yaml.DumpOptions = {
         indent: 2, // 缩进空格数
         noArrayIndent: false, // 数组是否缩进
         skipInvalid: false, // 跳过无效类型
@@ -56,5 +42,58 @@ export function dump(form: IForm): string {
         quotingType: '"', // 引号类型
         forceQuotes: false, // 强制引号
         // replacer: (k, v) => {}, // 回调函数
-    })
+    },
+): string {
+    Object.assign(defaultOptions, options);
+    return yaml.dump(obj, defaultOptions);
+}
+
+/* 序列化 IAL */
+export function dumpIAL(attrs: IMetadata[]): string {
+    const ial: Record<string, any> = {};
+    attrs.forEach(attr => {
+        ial[attr._key] = (() => {
+            switch (attr.key) {
+                case "created":
+                case "updated":
+                    return moment(attr.value, "YYYYMMDDHHmmss").toDate();
+                case "tags":
+                case "alias":
+                    return tokenSplit(attr.value);
+                default:
+                    return attr.value;
+            }
+        })();
+    });
+    return dump(ial, {
+        sortKeys: (k1, k2) => {
+            const l1 = attrs.findIndex(attr => attr._key === k1);
+            const l2 = attrs.findIndex(attr => attr._key === k2);
+            return l1 - l2;
+        },
+    });
+}
+
+/* 序列化 Form */
+export function dumpForm(form: IForm): string {
+    const ial: Record<string, any> = {
+        title: form.basics.title,
+        id: form.others.id,
+        date: moment(form.basics.created, "YYYYMMDDHHmmss").toDate(),
+        lastmod: moment(form.basics.updated, "YYYYMMDDHHmmss").toDate(),
+        // date: moment(form.basics.created, "YYYYMMDDHHmmss").utc().format(),
+        // lastmod: moment(form.basics.updated, "YYYYMMDDHHmmss").utc().format(),
+    };
+
+    if (form.basics.name) ial.name = form.basics.name;
+    if (form.basics.memo) ial.memo = form.basics.memo;
+    if (form.basics.tags.length > 0) ial.tags = form.basics.tags;
+    if (form.basics.alias.length > 0) ial.alias = form.basics.alias;
+
+    form.customs.reduce((ial, custom) => {
+        ial[custom.key] = custom.value;
+        return ial;
+    }, ial);
+
+    return dump(ial);
 }