+++ public/CHANGELOG.md
@@ -7,6 +7,7 @@
 - 调整响应式布局样式 | Adjust responsive layout style.
 - 添加导出标签页 | Add export tab.
 - 启用 Arco 组件的国际化功能 | Enable internationalization of Arco components.
+- 实现导出标签页 | Implement export tab.
 
 ## v0.1.1 | 2023-03-23